requirements.txt
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.21.0
WTForms==3.0.1
